Reorganize configuration classes
