+++ LyricAnimator/Animator.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using LyricAnimator.Configuration;
 using SkiaSharp;
 
 namespace LyricAnimator
