+++ LyricAnimator/Configuration/Lyric.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace LyricAnimator.Configuration
+{
+    internal class Lyric
+    {
+        public TimeSpan StartTime { get; set; }
+        public TimeSpan EndTime { get; set; }
+        public List<string> Lines { get; set; }
+        public int VerseNumber { get; set; }
+    }
+}
