+++ LyricAnimator/MainWindow.xaml.cs
@@ -7,6 +7,7 @@ using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Input;
+using LyricAnimator.Configuration;
 using Newtonsoft.Json;
 
 namespace LyricAnimator
